move wifi methods inside main
